adding makeIndex and makeIndexType back to tracks.ts as exported functions for plugins to use
