+++ plugins/gff3/src/index.ts
@@ -4,6 +4,7 @@ import Plugin from '@jbrowse/core/Plugin'
 import { configSchema as gff3TabixAdapterConfigSchema } from './Gff3TabixAdapter'
 import AdapterGuessType from '@jbrowse/core/pluggableElementTypes/AdapterGuessType'
 import { FileLocation } from '@jbrowse/core/util/types'
+import { makeIndex, makeIndexType } from '@jbrowse/core/util/tracks'
 
 export default class extends Plugin {
   name = 'GFF3TabixPlugin'
